Add GitHub workflow to auto update (rebuild & publish) containers
`php-dev` is a "fat container" including many tools, too many actually. Furthermore, it's a development environment, no production service for which it would be crucial to receive updates immediately. So, we simply rebuild once a week. We can be rather certain that there are updates by then, so don't bother to check, just always rebuild.